+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - **CHANGED**
   - Update client to version v1.0.89
   ([PR #130](https://github.com/cycloidio/cycloid-cli/pull/130))
+  - Add the ability to specify canonical for credential create
+  ([PR #130](https://github.com/cycloidio/cycloid-cli/pull/130))
 
 ## [v1.0.88] _2022-02-10_
 - **CHANGED**